+++ src/pytask_stata/shared.py
@@ -3,10 +3,10 @@
 from __future__ import annotations
 
 import sys
+from collections.abc import Iterable
+from collections.abc import Sequence
 from typing import TYPE_CHECKING
 from typing import Any
-from typing import Iterable
-from typing import Sequence
 
 if TYPE_CHECKING:
     from pathlib import Path