+++ tests/test_execute.py
@@ -18,7 +18,6 @@ from pytask_stata.execute import pytask_execute_task_setup
 from tests.conftest import needs_stata
 
 
-@pytest.mark.unit
 @pytest.mark.parametrize(
     ("stata", "expectation"),
     [(executable, does_not_raise()) for executable in STATA_COMMANDS]
@@ -31,7 +30,7 @@ def test_pytask_execute_task_setup_raise_error(stata, platform, expectation):
     task = Task(
         base_name="task_example",
         path=Path(),
-        function=None,
+        function=lambda: None,  # ty: ignore[invalid-argument-type]
         markers=[Mark("stata", (), {})],
     )
 
@@ -42,7 +41,6 @@ def test_pytask_execute_task_setup_raise_error(stata, platform, expectation):
 
 
 @needs_stata
-@pytest.mark.end_to_end
 def test_run_do_file(runner, tmp_path):
     task_source = """
     import pytask
@@ -72,7 +70,6 @@ def test_run_do_file(runner, tmp_path):
 
 
 @needs_stata
-@pytest.mark.end_to_end
 def test_run_do_file_w_task_decorator(runner, tmp_path):
     task_source = """
     import pytask
@@ -102,7 +99,6 @@ def test_run_do_file_w_task_decorator(runner, tmp_path):
         assert tmp_path.joinpath("script.log").exists()
 
 
-@pytest.mark.end_to_end
 def test_raise_error_if_stata_is_not_found(tmp_path, monkeypatch):
     task_source = """
     from pytask import mark, task
@@ -124,11 +120,10 @@ def test_raise_error_if_stata_is_not_found(tmp_path, monkeypatch):
     session = build(paths=tmp_path)
 
     assert session.exit_code == ExitCode.FAILED
-    assert isinstance(session.execution_reports[0].exc_info[1], RuntimeError)
+    assert isinstance(session.execution_reports[0].exc_info[1], RuntimeError)  # ty: ignore[non-subscriptable]
 
 
 @needs_stata
-@pytest.mark.end_to_end
 def test_run_do_file_w_wrong_cmd_option(runner, tmp_path):
     """Apparently, Stata simply discards wrong cmd options."""
     task_source = """
@@ -154,7 +149,6 @@ def test_run_do_file_w_wrong_cmd_option(runner, tmp_path):
 
 
 @needs_stata
-@pytest.mark.end_to_end
 def test_run_do_file_by_passing_path(runner, tmp_path):
     """Replicates example under "Command Line Arguments" in Readme."""
     task_source = """
@@ -180,7 +174,6 @@ def test_run_do_file_by_passing_path(runner, tmp_path):
 
 
 @needs_stata
-@pytest.mark.end_to_end
 def test_run_do_file_fails_with_multiple_marks(runner, tmp_path):
     task_source = """
     import pytask
@@ -201,7 +194,6 @@ def test_run_do_file_fails_with_multiple_marks(runner, tmp_path):
 
 
 @needs_stata
-@pytest.mark.end_to_end
 def test_with_task_without_path(runner, tmp_path):
     task_source = """
     import pytask